- round 3 of css refactoring - rating list positioning fixed for mobile - css transitions
