+++ frontend/src/Components/CommentComponent.module.scss
@@ -1,0 +1,133 @@
+.comment {
+    margin: 20px 0;
+    .content>div {
+        font-size: 16px;
+        line-height: 1.5em;
+        color: var(--fg);
+        overflow-wrap: anywhere;
+        @media only screen and (max-width: 600px) {
+            font-size: 14px;
+        }
+    }
+    img {
+         max-width: 100%;
+         max-height: 500px;
+    }
+}
+
+.signature {
+    display: flex;
+    align-items: center;
+    gap: 4px;
+    font-size: 14px;
+    margin-bottom: 4px;
+    color: var(--fgSoftest);
+    a {
+        text-decoration: none;
+        color: var(--fgSoftest);
+        &:hover {
+            color: var(--fgSofter);
+        }
+    }
+    .username {
+        color: var(--fgMedium);
+        svg{ fill: var(--fgSoftest); }
+        &:hover {
+            color: var(--primary);
+            svg{  fill: var(--primary); }
+        }
+    }
+}
+
+.controls {
+    display: flex;
+    margin-top: 4px;
+    margin-left: -3px;
+    align-items: center;
+
+    .control {
+        margin-right: 15px;
+        color: var(--fgSofter);
+        display: flex;
+        align-items: center;
+    }
+
+    button, a {
+        height: 24px;
+        background: none;
+        margin: 0;
+        padding: 0;
+        border: none;
+        display: flex;
+        align-items: center;
+        font-size: 13px;
+        fill: var(--fgSoftest);
+        color: var(--fgSoftest);
+        text-decoration: none;
+
+        font-weight: bold;
+
+        &.active {
+            fill: var(--primary);
+            font-weight: bold;
+        }
+    }
+    a {
+         margin: 0 2px;
+        &.active {
+            color: var(--primary);
+        }
+    }
+    button:disabled {
+          fill: var(--fgGhost);
+          cursor: default;
+    }
+    .label {
+         margin-left: 5px;
+    }
+}
+
+.answers {
+    margin-top: 32px;
+    margin-bottom: 32px;
+    padding-left: 8px;
+    border-left: 1px solid var(--fgAlmostInvisible);
+
+    @media only screen and (min-width: 990px) {
+        padding-left: 20px;
+        margin-left: 8px;
+    }
+}
+
+
+.unreadOnly .comment:not(.isNew) > .body {
+    opacity: 0.4;
+    transition: opacity 500ms;
+}
+
+.unreadOnly .comment:not(.isNew) > .body:hover {
+    opacity: inherit;
+    transition: opacity 500ms;
+}
+
+.isNew {
+    border-left: 2px solid var(--primary);
+    margin-left: -8px;
+    padding-left: 8px;
+
+    @media only screen and (min-width: 990px) {
+        margin-left: -21px;
+        padding-left: 21px;
+    }
+}
+
+
+/* highlight only top level of unread comments */
+.isNew .isNew {
+    border-left: 1px solid var(--fgAlmostInvisible);
+}
+
+.newComment {
+    color: var(--primary);
+    font-style: italic;
+}
