+++ frontend/src/Components/CreateCommentComponent.tsx
@@ -2,14 +2,14 @@ import {CommentInfo, PostLinkInfo} from '../Types/PostInfo';
 import React, {useEffect, useRef, useState} from 'react';
 import styles from './CreateCommentComponent.module.scss';
 import postStyles from '../Pages/CreatePostPage.module.css';
-import commentStyles from './CommentComponent.module.css';
+import commentStyles from './CommentComponent.module.scss';
 import {ReactComponent as IronyIcon} from '../Assets/irony.svg';
 import {ReactComponent as ImageIcon} from '../Assets/image.svg';
 import {ReactComponent as LinkIcon} from '../Assets/link.svg';
 import {ReactComponent as QuoteIcon} from '../Assets/quote.svg';
 import {ReactComponent as SendIcon} from '../Assets/send.svg';
 import ContentComponent from "./ContentComponent";
-import classNames from "classnames";
+import classNames from 'classnames';
 import MediaUploader from './MediaUploader';
 
 interface CreateCommentProps {
@@ -25,7 +25,7 @@ export default function CreateCommentComponent(props: CreateCommentProps) {
     const answerRef = useRef<HTMLTextAreaElement>(null);
 
     const pronoun = props?.comment?.author?.gender == 1 ? 'ему' : props?.comment?.author?.gender==2 ? 'ей' : '';
-    const [placeholderText, setPlaceholderText] = useState<string>(props.comment ? `Ваш ответ ${pronoun}` : '');
+    const [placeholderText] = useState<string>(props.comment ? `Ваш ответ ${pronoun}` : '');
     const [answerText, setAnswerText] = useState<string>( '');
     const [isPosting, setPosting] = useState(false);
     const [previewing, setPreviewing] = useState<string | null>(null);
@@ -207,7 +207,7 @@ export default function CreateCommentComponent(props: CreateCommentProps) {
                 :  <div className={classNames(commentStyles.content, styles.preview, postStyles.preview)} onClick={handlePreview}><ContentComponent content={previewing} /></div>
             }
             <div className={styles.final}>
-                <button disabled={isPosting || !answerText} className={styles.buttonPreview} onClick={handlePreview}>{(previewing === null) ? "Превью" : "Редактор"}</button>
+                <button disabled={isPosting || !answerText} className={styles.buttonPreview} onClick={handlePreview}>{(previewing === null) ? 'Превью' : 'Редактор'}</button>
                 <button disabled={isPosting || !answerText} className={styles.buttonSend} onClick={handleAnswer}><SendIcon /></button>
                 {mediaUploaderOpen && <MediaUploader onSuccess={handleMediaUpload} onCancel={handleMediaUploadCancel} />}
             </div>
