+++ frontend/src/Pages/PostPage.module.css
@@ -6,6 +6,7 @@
 
 .feed {
 
+    transition-duration: 0s;
 }
 
 .comments {
@@ -16,6 +17,8 @@
     margin: 30px 0;
     border-radius: 6px;
     font-size: 14px;
+    transition-duration: 0s;
+    color: var(--fgSoftest);
 }
 
 .postButtons a {
